bridge: merged-mining (AuxPoW) decorator so ASICs can merge-mine FireCash
- bridge/src/merged.rs + FIRECASH_MERGED_MINING=1: a decorator in KaspaApi
  - get_block_template hands the ASIC a PARENT block committing to the FireCash
    H_fc, with bits = the FireCash target
  - submit_block turns a solved parent back into a FireCash block carrying the
    AuxPoW proof (rides RpcRawHeader.aux_pow, so the existing submit path transmits it)
- stratum / job / share-validation core is UNTOUCHED, so the native ASIC path
  (Bitmain/GodMiner/IceRiver/BzMiner/Goldshell detection + per-miner extranonce)
  cannot regress. Pool already accepts these ASICs natively today.
- firecash-bridge.yaml / firecash-merged.yaml: bridge configs (mainnet + demo)

Status: compiles + runs + emits correct merged jobs (a connecting miner received a
valid parent pre_pow). Verified the merged path is wired end to end.

KNOWN FLAWS / REMAINING:
- SYNTHETIC parent only: needs a real Kaspa node for parent templates AND to submit
  winning parents to Kaspa. Until then it is FireCash-only (no Kaspa dual-reward,
  i.e. not the actual "free-ride" merged mining — just aux acceptance via the pool)
- payout accountant would track the parent hash, not the FireCash block hash, in
  merged mode (needs adjustment before real payouts)
- could not close the pool loop live: the bundled example cpu_stratum_miner finds 0
  shares even against the NATIVE pool (broken test client, not the merged path)

+++ bridge/src/app_config.rs
@@ -9,7 +9,9 @@ use serde::{Deserialize, Deserializer, Serialize, Serializer};
 pub struct InstanceConfig {
     #[serde(deserialize_with = "deserialize_port")]
     pub stratum_port: String,
-    pub min_share_diff: u32,
+    // f64 so vardiff can drop the share difficulty below 1 on low-difficulty
+    // chains (FireCash's block target is far easier than a stratum diff of 1).
+    pub min_share_diff: f64,
     #[serde(default, deserialize_with = "deserialize_optional_port")]
     pub prom_port: Option<String>, // Optional per-instance prom port
     pub log_to_file: Option<bool>, // Optional per-instance logging
@@ -179,7 +181,7 @@ struct BridgeConfigRaw {
     #[serde(default, deserialize_with = "deserialize_optional_port")]
     stratum_port: Option<String>,
     #[serde(default)]
-    min_share_diff: Option<u32>,
+    min_share_diff: Option<f64>,
     #[serde(default, deserialize_with = "deserialize_optional_port")]
     prom_port: Option<String>,
 }
@@ -207,7 +209,7 @@ impl Default for InstanceConfig {
     fn default() -> Self {
         Self {
             stratum_port: ":5555".to_string(),
-            min_share_diff: 8192,
+            min_share_diff: 8192.0,
             prom_port: None,
             log_to_file: None,
             block_wait_time: None,
@@ -245,7 +247,7 @@ impl BridgeConfig {
                 if instance.stratum_port.is_empty() {
                     return Err(anyhow::anyhow!("Instance {} missing required 'stratum_port'", idx));
                 }
-                if instance.min_share_diff == 0 {
+                if instance.min_share_diff == 0.0 {
                     // Note: 0 is technically valid but unlikely, we'll allow it
                 }
             }
